improved OAuth2Controller test
